requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
idna==3.3
mysqlclient==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pep8==1.7.1
pip-upgrader==1.4.15
pycodestyle==2.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
toml==0.10.2
tzdata==2022.2
urllib3==1.26.11
googletrans==3.1.0a0